[Silver III] Title: 주유소, Time: 376 ms, Memory: 35632 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.java
@@ -33,7 +33,7 @@ public class Main {
             }
 
             if (nextPos == -1) {
-                int distSum = 0;
+                long distSum = 0;
                 for (int i = currentPos; i < N - 1; i++) {
                     distSum += dist[i];
                 }
@@ -41,7 +41,7 @@ public class Main {
                 break;
             }
 
-            int distSum = 0;
+            long distSum = 0;
             for (int i = currentPos; i < nextPos; i++) {
                 distSum += dist[i];
             }
